[MODIFY]: move getContentSize() method from BPPage class to BPNode
